Rename addTransaction,printTransactions to addTransactionprintTransactions
This script defines two functions: addTransaction, which can be used to add a new transaction to the list, and printTransactions, which can be used to print the list of transactions. You can use these functions to build a payment app similar to Venmo. For example, you could use the addTransaction function to process payments between users, and the printTransactions function to display a list of transactions to the user.