Create payment methods
This script defines two functions: addPaymentMethod, which can be used to add a new payment method to the list, and printPaymentMethods, which can be used to print the list of payment methods. You can use these functions to allow users to select multiple payment methods in your payment app. For example, you could use the addPaymentMethod function to process user input when they add a new payment method, and the printPaymentMethods function to display the list of payment methods to the user.