Create customer support
This script defines an object to store the customer support resources, including a contact email address and a list of frequently asked questions (FAQs). It also defines two functions: printContact, which can be used to print the customer support contact information, and printFAQs, which can be used to print the list of FAQs. You can use these functions to provide access to customer support resources in your payment app. For example, you could use the printContact function to display the customer support contact information to the user, and the printFAQs function to display the list of FAQs when the user needs help with the app.