Create QR CODE transactions
This script defines two functions: generateQR, which can be used to generate a QR code for a transaction, and scanQR, which can be used to scan a QR code and process the transaction. You can use these functions to send and receive transactions using QR codes in your payment app. For example, you could use the generateQR function to create a QR code for a transaction when the user wants to send a payment, and the scanQR function to process a payment when the user scans a QR code they received from another user.